Make DataTypeService an interface
While at it improve the internal implementation of the default
DataTypeService.

+++ core/data/src/main/java/imagej/data/types/DataType12BitUnsignedInteger.java
@@ -123,7 +123,7 @@ public class DataType12BitUnsignedInteger extends AbstractContextual implements
 		short v = dest.get();
 		if (v < 0) dest.set((short) 0);
 		else if (v > 0xfff) dest.set((short) 0xfff);
-		else dest.set((short) v);
+		else dest.set(v);
 	}
 
 }
